fix: Server util functions on separate file
